Alpine and Muslcc llvm compile not working
This approach is here to solely show muslcc and alpine aren't ideal environments to build LLVM via musl-libc anymore due to the lack of execinfo.h (provided by libexecinfo in muslcc and alpine linux before version 3.17 was introduced but it is now discontinued), which even if installed in muslcc's docker container fails to get found when llvm's project is attempted to be built.

Additionally, no set of LLVM flags that previous llvm-dev-email listings/pages and any forum pages have used for this purpose are working - hence the decision to forgo compiling libc++ with musl libc.