Transition to Zig for C and C++ static compiles
This resolves the need for creating a custom toolchain to compile llvm with musl-libc as the target. Ultimately allowing any produced binary to run on any system without reliance on glibc or libstdc++, instead opting for musl-libc and libc++ that're MIT and Apache License v2.0 with LLVM Exceptions.

Still a linker error from a --dependency-file arg, but will fix with followup commits.