Zig/C/C++ Cross-platform toolchain achieved
From this point onwards, Zig's static linkage against musl-libc and libc++ for the Cobralyth compiler can be used moving forwards on linux.
If supporting other platforms, we'd just need to create the binaries for said platforms as well. Though for now, just targeting linux OS for x86_64, with plans for ARM support for RPI4 once the language matures.